Remove extraneous options in mask_and_scale
This commit removes extraneous scaling options in the mask_and_scale() functions of the obs and model classes in driver.py. The removed scaling options can be accomplished with `unit_scale` and `unit_scale_method`. The options were not included in the documentation and so do not need to be removed from there.